requirements.txt
pandas==2.1.4
azure-cognitiveservices-speech==1.31.0
python-dotenv==1.0.0
sounddevice==0.4.6
numpy==1.26.4
pathlib==1.0.1
logging==0.5.1.2
ttkthemes==3.2.2
pillow==10.2.0